Restructure lighthouse files
