+++ ui/lib/lighthouse/prompts.ts
