+++ ui/lib/lighthouse/tools/roles.ts
