+++ ui/lib/lighthouse/utils.ts
